requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
gunicorn>=20.1.0
pymysql>=1.1.0
cryptography>=41.0.0
google-genai
